Changed argument timeForceLost, now it has type of uint and assumes milliseconds (was float in seconds) in ParticleSystemSceneNode.CreateGravityAffector() method.

+++ trunk/source/ParticleSystemSceneNode.h
@@ -71,7 +71,7 @@ public:
 	ParticleFadeOutAffector^ CreateFadeOutParticleAffector(Video::Color^ targetColor);
 	ParticleFadeOutAffector^ CreateFadeOutParticleAffector();
 
-	ParticleGravityAffector^ CreateGravityAffector(Vector3Df^ gravity, float timeForceLost);
+	ParticleGravityAffector^ CreateGravityAffector(Vector3Df^ gravity, unsigned int timeForceLost);
 	ParticleGravityAffector^ CreateGravityAffector(Vector3Df^ gravity);
 	ParticleGravityAffector^ CreateGravityAffector();
